+++ MSA_UI.lua
@@ -6,10 +6,6 @@ MSA.UI = UI;
 -- Holds GUI frame load configuration details
 UI.LoadUI = function()
 
-    if not MSA.Crafting.Profs then
-        MSA.Crafting.Establish_Spells()
-    end
-
     -- Checkbox to be placed on inscription window
     if not UI.MSA_checkbox then
 
@@ -31,6 +27,7 @@ UI.LoadUI = function()
         -- Change the setting wether enabled or not
         UI.MSA_checkbox:SetScript ( "OnClick" , function( self )
             MSA_save.non_stop = self:GetChecked()
+            UI.MSA_checkbox.value = MSA_save.non_stop
         end)
 
         -- Tooltip
@@ -54,6 +51,114 @@ UI.LoadUI = function()
 
 end
 
+----------------------------
+------ UI DEVELOPMENT ------
+---------- TIMER -----------
+----------------------------
+
+-- Deploy it
+UI.Deploy_Timer_UI = function()
+
+    if not UI.CT_Core_Frame then
+        -- Core frame
+        UI.CT_Core_Frame = CreateFrame( "Frame" , "CT_Core_Frame" , UIParent , BackdropTemplateMixin and "BackdropTemplate" );
+        UI.CT_Core_Frame:SetBackdrop ( {
+            bgFile = "Interface\\DialogFrame\\UI-DialogBox-Background" ,
+            edgeFile = "Interface\\DialogFrame\\UI-DialogBox-Border",
+            tile = true,
+            tileSize = 32,
+            edgeSize = 8,
+            insets = { left = 2 , right = 2 , top = 3 , bottom = 2 }
+         })
+        UI.CT_Core_Frame:SetPoint ( MSA_save.pos[1] , UIParent , MSA_save.pos[2] , MSA_save.pos[3] , MSA_save.pos[4] );
+        UI.CT_Core_Frame:SetSize( 300 , 120 );
+        UI.CT_Core_Frame:EnableMouse ( true );
+        UI.CT_Core_Frame:SetToplevel ( true );
+        UI.CT_Core_Frame:EnableMouse ( true );
+        UI.CT_Core_Frame:SetMovable ( true );
+        UI.CT_Core_Frame:RegisterForDrag ( "LeftButton" );
+        UI.CT_Core_Frame:SetScript ( "OnDragStart" , UI.CT_Core_Frame.StartMoving )
+        UI.CT_Core_Frame:SetScript ( "OnDragStop" , function( self )
+            self:StopMovingOrSizing()
+            local pos1 , _ , pos2 , x , y = self:GetPoint();
+            MSA_save.pos = { pos1 , pos2 , x , y };
+        end);
+        UI.CT_Core_Frame.value = 0              -- Will be placeholder of the seconds...
+        UI.CT_Core_Frame.On_Update_Timer = 0;   -- Control speed of on_update.
+        UI.CT_Core_Frame.craft_id = 0;          -- Keep tabs on the craft ID
+        UI.CT_Core_Frame.refresh_count = 0      -- Refresh the countdown every X crafts
+
+        -- Keep it closed at start
+        UI.CT_Core_Frame:Hide();
+
+        UI.CT_Core_Frame:SetScript( "OnShow" , function()
+            if C_TradeSkillUI.IsRecipeRepeating() then
+                if #CT.timer_table < 10 then
+                    MSA.UI.CT_Core_Frame.Countdown_Text:SetText( "Calculating..." )
+                elseif MSA.UI.CT_Core_Frame.value > 0 then
+                    MSA.UI.CT_Core_Frame.Countdown_Text:SetText( formatTime ( MSA.UI.CT_Core_Frame.value ) );
+                end
+            end
+        end)
+
+        -- Close Button
+        UI.CT_Core_Frame.CT_CloseButton = CreateFrame( "Button" , "CT_CloseButton" , UI.CT_Core_Frame , "UIPanelCloseButton");
+        UI.CT_Core_Frame.CT_CloseButton:SetPoint( "TOPRIGHT" , UI.CT_Core_Frame , "TOPRIGHT" , -4 , -4 );
+        UI.CT_Core_Frame.CT_CloseButton:SetSize ( 26 , 26 );
+
+        -- Font Strings Title and main
+        UI.CT_Core_Frame.Header_Text = UI.CT_Core_Frame:CreateFontString ( nil , "OVERLAY" , "GameFontWhiteTiny" );
+        UI.CT_Core_Frame.Header_Text:SetPoint ( "TOP" , UI.CT_Core_Frame, "TOP" , 0 , -12 );
+        UI.CT_Core_Frame.Header_Text:SetFont( STANDARD_TEXT_FONT , 14 , "BOLD");
+        UI.CT_Core_Frame.Header_Text:SetText ( "Time Remaining" );
+
+        UI.CT_Core_Frame.Countdown_Text = UI.CT_Core_Frame:CreateFontString ( nil , "OVERLAY" , "GameFontWhite" );
+        UI.CT_Core_Frame.Countdown_Text:SetPoint ( "TOP" , UI.CT_Core_Frame.Header_Text , "BOTTOM" , 0 , -5 );
+
+
+        --------------------------------
+        -------- TIMER SETTINGS --------
+        ---------- CHECKBOXES ----------
+
+        -- MSA_save.count_bags_Only  == nil then        -- Only applies when Salvaging
+        --     MSA_save.count_bags_Only = true
+        -- end
+
+        -- if MSA_save.always_show
+
+        UI.CT_Core_Frame.Always_Show_Checkbox = CreateFrame ( "CheckButton" , "Always_Show_Checkbox" , UI.CT_Core_Frame , "InterfaceOptionsCheckButtonTemplate" )
+        UI.CT_Core_Frame.Always_Show_Checkbox.value = MSA_save.always_show
+        UI.CT_Core_Frame.Always_Show_Checkbox:SetChecked ( UI.CT_Core_Frame.Always_Show_Checkbox.value )
+
+        -- Text to the right of checkbox
+        UI.CT_Core_Frame.Always_Show_Checkbox.Text = UI.CT_Core_Frame.Always_Show_Checkbox:CreateFontString ( nil , "OVERLAY" , "GameFontNormal" )
+        UI.CT_Core_Frame.Always_Show_Checkbox.Text:SetFont( STANDARD_TEXT_FONT , 12 , "BOLD");
+        UI.CT_Core_Frame.Always_Show_Checkbox.Text:SetText( "Always Show Timer When Crafting" )
+        UI.CT_Core_Frame.Always_Show_Checkbox.Text:SetPoint( "LEFT" , UI.CT_Core_Frame.Always_Show_Checkbox, "RIGHT" , 2 , 0 )
+
+        -- Normalize the click area of check button to length of the text
+        UI.CT_Core_Frame.Always_Show_Checkbox:SetHitRectInsets ( 0 , 0 - UI.CT_Core_Frame.Always_Show_Checkbox.Text:GetWidth() - 2 , 0 , 0 );
+
+        -- Ensures this is always to the left of the CreateAllButton, accounting for width of text also
+        UI.CT_Core_Frame.Always_Show_Checkbox:SetPoint ( "BOTTOMLEFT" , UI.CT_Core_Frame , "BOTTOMLEFT" , 10 , 10 )
+
+        -- Change the setting wether enabled or not
+        UI.CT_Core_Frame.Always_Show_Checkbox:SetScript ( "OnClick" , function( self )
+            MSA_save.always_show = self:GetChecked()
+            UI.MSA_checkbox.value = MSA_save.always_show
+            if not UI.MSA_checkbox.value then
+                print('If you wish to reopen the timer again, type `/msa timer' )
+            end
+        end)
+
+    end
+
+end
+
+-----------------
+--- UI LOGIC ----
+-----------------
+
 -- Method:          UI.Is_Supported_Profession ( int )
 -- What it Does:    Returns true if the profession has been added and configured by addon dev for mass crafting
 -- Purpose:         Addon will continue to expand usage.
@@ -68,6 +173,7 @@ UI.Is_Supported_Profession = function ( profession_id )
         [197] = "Tailoring",
         [202] = "Engineering",
         [333] = 'Enchanting',
+        [393] = "Skinning",
         [755] = "Jewelcrafting",
         [773] = "Inscription"
     }